@@ -99,7 +99,7 @@
     "source-map-explorer": "^2.5.3",
     "start-server-and-test": "^1.15.3",
     "tsconfig-paths-webpack-plugin": "^4.0.1",
-    "typescript": "^4.9.5",
+    "typescript": "^5.2.0",
     "util": "^0.12.5",
     "vitest": "^2.1.2",
     "vitest-canvas-mock": "^0.3.3",
