@@ -106,7 +106,7 @@
     "sprintf-js": "^1.1.3",
     "styletron-react": "^6.1.0",
     "typed-signals": "^2.5.0",
-    "typescript": "^4.9.5",
+    "typescript": "^5.2.0",
     "ua-parser-js": "^1.0.39",
     "unist-util-visit": "^4.1.2",
     "uuid": "^9.0.0",
